+++ test/shell_classify/classifiers/perl_test.clj
@@ -1,7 +1,7 @@
-(ns shell-shape-classify.classifiers.perl-test
+(ns shell-classify.classifiers.perl-test
   "Perl stdlib effect classifier (P5.4 / v0.16.0)."
   (:require [clojure.test :refer [deftest is testing]]
-            [shell-shape-classify.classify :as cls]
+            [shell-classify.classify :as cls]
             [shell-shape.dialect.perl :as perl]))
 
 (defn- effects-of [src]